requirements.txt
Django==1.9.8
django-bootstrap3==6.2.2
django-html5==1.0.0